@@ -1,5 +1,7 @@
+from __future__ import annotations
+
 from functools import lru_cache
-from typing import Any, Dict, Optional
+from typing import Any
 
 from pydantic import Field, ValidationError, model_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
@@ -9,13 +11,13 @@ class MeilisearchConfig(BaseSettings):
     meili_http_addr: str
     meili_https_url: bool = False
     meilisearch_url: str = "http://localhost:7700"
-    meilisearch_api_key: Optional[str] = Field(None, validation_alias="MEILI_MASTER_KEY")
+    meilisearch_api_key: str | None = Field(None, validation_alias="MEILI_MASTER_KEY")
     model_config = SettingsConfigDict(
         env_file=".env", env_file_encoding="utf-8", case_sensitive=False, extra="allow"
     )
 
     @model_validator(mode="before")
-    def set_url(cls, values: Dict[str, Any]) -> Dict[str, Any]:
+    def set_url(cls, values: dict[str, Any]) -> dict[str, Any]:
         if not values.get("meili_http_addr"):
             raise ValidationError("A MEILI_HTTP_ADDR value is required", MeilisearchConfig)
 
