@@ -1,5 +1,7 @@
+from __future__ import annotations
+
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Union
+from typing import Any
 
 from camel_converter.pydantic_base import CamelBase
 from meilisearch_python_sdk.models.client import Key
@@ -10,6 +12,6 @@ class TenantToken(CamelBase):
 
 
 class TenantTokenSettings(CamelBase):
-    search_rules: Union[Dict[str, Any], List[str]]
+    search_rules: dict[str, Any] | list[str]
     api_key: Key
-    expires_at: Optional[datetime] = None
+    expires_at: datetime | None = None
