@@ -72,7 +72,9 @@ async def fastapi_test_client():
     app.include_router(api_router)
 
     async with AsyncClient(
-        transport=ASGITransport(app=app), base_url="http://test/", follow_redirects=True
+        transport=ASGITransport(app=app),  # type: ignore
+        base_url="http://test/",
+        follow_redirects=True,
     ) as ac:
         yield ac
 
